fix(lint): clippy::float_cmp + redundant cast (CI -D warnings)
CI's `cargo clippy --workspace --all-targets -- -D warnings` (clippy 1.95)
failed two lints that local dev-profile clippy didn't escalate:

1. butteraugli/src/lib.rs:1131 — `assert_eq!(res.max_norm(), 1.234)` triggers
   `clippy::float_cmp`. Switch to `to_bits()` comparison: max_norm() is just
   a getter, must round-trip exactly.

2. butteraugli-bench/examples/score_parity.rs:208 — `rust_score as f64` is
   a redundant cast (rust_score is already f64).

+++ butteraugli-bench/examples/score_parity.rs
@@ -205,7 +205,7 @@ fn main() {
             #[cfg(not(has_cpp_butteraugli))]
             let cpp_score = f64::NAN;
 
-            let abs = (rust_score as f64 - cpp_score).abs();
+            let abs = (rust_score - cpp_score).abs();
             // Avoid div-by-zero on identical-image cases.
             let rel = if cpp_score.abs() > 1e-9 {
                 abs / cpp_score.abs() * 100.0
